requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==4.2
djangorestframework==3.15.1
drf-yasg==1.21.7
Faker==25.9.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
inflection==0.5.1
packaging==24.1
protobuf==4.25.3
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wrapt==1.16.0
zipp==3.19.2
django-environ
pytest
selenium
webdriver-manager
pytest
pytest-django
pytest-xdist